+++ packages/coinlib-bitcoin/src/BaseBitcoinPayments.ts
@@ -151,151 +151,151 @@ export abstract class BaseBitcoinPayments<Config extends BaseBitcoinPaymentsConf
   }
 
 
-  private validatePsbtOutput(output: BitcoinishTxOutput, psbtOutput: bitcoin.PsbtTxOutput, i: number) {
-    if (output.address !== psbtOutput.address) {
-      throw new Error(`Invalid tx: psbt output ${i} address (${psbtOutput.address}) doesn't match expected address ${output.address}`)
-    }
-    const value = this.toMainDenomination(psbtOutput.value)
-    if (output.value !== value) {
-      throw new Error(`Invalid tx: psbt output ${i} value (${value}) doesn't match expected value (${output.value})`)
-    }
-  }
+  // private validatePsbtOutput(output: BitcoinishTxOutput, psbtOutput: bitcoin.PsbtTxOutput, i: number) {
+  //   if (output.address !== psbtOutput.address) {
+  //     throw new Error(`Invalid tx: psbt output ${i} address (${psbtOutput.address}) doesn't match expected address ${output.address}`)
+  //   }
+  //   const value = this.toMainDenomination(psbtOutput.value)
+  //   if (output.value !== value) {
+  //     throw new Error(`Invalid tx: psbt output ${i} value (${value}) doesn't match expected value (${output.value})`)
+  //   }
+  // }
 
-  private validatePsbtInput(input: UtxoInfo, psbtInput: bitcoin.PsbtTxInput, i: number) {
-    // bitcoinjs psbt input hash buffer is reversed
-    const hash = Buffer.from(psbtInput.hash).reverse().toString('hex')
-    if (input.txid !== hash) {
-      throw new Error(`Invalid tx: psbt input ${i} hash (${hash}) doesn't match expected txid (${input.txid})`)
-    }
-    if (input.vout !== psbtInput.index) {
-      throw new Error(
-        `Invalid tx: psbt input ${i} index (${psbtInput.index}) doesn't match expected vout (${input.vout})`
-      )
-    }
-  }
+  // private validatePsbtInput(input: UtxoInfo, psbtInput: bitcoin.PsbtTxInput, i: number) {
+  //   // bitcoinjs psbt input hash buffer is reversed
+  //   const hash = Buffer.from(psbtInput.hash).reverse().toString('hex')
+  //   if (input.txid !== hash) {
+  //     throw new Error(`Invalid tx: psbt input ${i} hash (${hash}) doesn't match expected txid (${input.txid})`)
+  //   }
+  //   if (input.vout !== psbtInput.index) {
+  //     throw new Error(
+  //       `Invalid tx: psbt input ${i} index (${psbtInput.index}) doesn't match expected vout (${input.vout})`
+  //     )
+  //   }
+  // }
 
-  /**
-   * Assert that a psbt is equivalent to the provided unsigned tx. Used to check a psbt actually
-   * reflects the expected transaction before signing.
-   */
-  validatePsbt(tx: BitcoinUnsignedTransaction, psbt: bitcoin.Psbt) {
-    const { data, externalOutputs, inputUtxos } = tx
-    const psbtOutputs = psbt.txOutputs
-    const psbtInputs = psbt.txInputs
+  // /**
+  //  * Assert that a psbt is equivalent to the provided unsigned tx. Used to check a psbt actually
+  //  * reflects the expected transaction before signing.
+  //  */
+  // validatePsbt(tx: BitcoinUnsignedTransaction, psbt: bitcoin.Psbt) {
+  //   const { data, externalOutputs, inputUtxos } = tx
+  //   const psbtOutputs = psbt.txOutputs
+  //   const psbtInputs = psbt.txInputs
 
-    if (!inputUtxos) {
-      throw new Error('Invalid tx: Missing inputUtxos')
-    }
-    if (!data.inputs) {
-      throw new Error('Invalid tx: Missing data.inputs')
-    }
-    if (!externalOutputs) {
-      throw new Error('Invalid tx: Missing externalOutputs')
-    }
-    if (!data.externalOutputs) {
-      throw new Error('Invalid tx: Missing data.externalOutputs')
-    }
-    if (!data.changeOutputs) {
-      throw new Error('Invalid tx: Missing data.changeOutputs')
-    }
-    if (!data.externalOutputTotal) {
-      throw new Error('Invalid tx: Missing data.externalOutputTotal')
-    }
-    if (!data.change) {
-      throw new Error('Invalid tx: Missing data.change')
-    }
+  //   if (!inputUtxos) {
+  //     throw new Error('Invalid tx: Missing inputUtxos')
+  //   }
+  //   if (!data.inputs) {
+  //     throw new Error('Invalid tx: Missing data.inputs')
+  //   }
+  //   if (!externalOutputs) {
+  //     throw new Error('Invalid tx: Missing externalOutputs')
+  //   }
+  //   if (!data.externalOutputs) {
+  //     throw new Error('Invalid tx: Missing data.externalOutputs')
+  //   }
+  //   if (!data.changeOutputs) {
+  //     throw new Error('Invalid tx: Missing data.changeOutputs')
+  //   }
+  //   if (!data.externalOutputTotal) {
+  //     throw new Error('Invalid tx: Missing data.externalOutputTotal')
+  //   }
+  //   if (!data.change) {
+  //     throw new Error('Invalid tx: Missing data.change')
+  //   }
 
-    // Check inputs
+  //   // Check inputs
 
-    if (inputUtxos.length !== data.inputs.length) {
-      throw new Error(
-        `Invalid tx: inputUtxos length (${psbtInputs.length}) doesn't match data.inputs length (${data.inputs.length})`
-      )
-    }
-    if (psbtInputs.length !== data.inputs.length) {
-      throw new Error(
-        `Invalid tx: psbt inputs length (${psbtInputs.length}) doesn't match data.inputs length (${data.inputs.length})`
-      )
-    }
+  //   if (inputUtxos.length !== data.inputs.length) {
+  //     throw new Error(
+  //       `Invalid tx: inputUtxos length (${psbtInputs.length}) doesn't match data.inputs length (${data.inputs.length})`
+  //     )
+  //   }
+  //   if (psbtInputs.length !== data.inputs.length) {
+  //     throw new Error(
+  //       `Invalid tx: psbt inputs length (${psbtInputs.length}) doesn't match data.inputs length (${data.inputs.length})`
+  //     )
+  //   }
 
-    let inputTotal = new BigNumber(0)
+  //   let inputTotal = new BigNumber(0)
 
-    // Safe to assume inputs are consistently ordered
-    for (let i = 0; i < psbtInputs.length; i++) {
-      const psbtInput = psbtInputs[i]
-      this.validatePsbtInput(inputUtxos[i], psbtInput, i)
-      this.validatePsbtInput(data.inputs[i], psbtInput, i)
-      inputTotal = inputTotal.plus(data.inputs[i].value)
-    }
+  //   // Safe to assume inputs are consistently ordered
+  //   for (let i = 0; i < psbtInputs.length; i++) {
+  //     const psbtInput = psbtInputs[i]
+  //     this.validatePsbtInput(inputUtxos[i], psbtInput, i)
+  //     this.validatePsbtInput(data.inputs[i], psbtInput, i)
+  //     inputTotal = inputTotal.plus(data.inputs[i].value)
+  //   }
 
-    // Check outputs
+  //   // Check outputs
 
-    if (externalOutputs.length !== data.externalOutputs.length) {
-      throw new Error(
-        `Invalid tx: externalOutputs length (${externalOutputs.length}) doesn't match data.externalOutputs length (${data.externalOutputs.length})`
-      )
-    }
-    const expectedOutputCount = data.externalOutputs.length + data.changeOutputs.length
-    if (psbtOutputs.length !== expectedOutputCount) {
-      throw new Error(
-        `Invalid tx: psbt outputs length (${psbtOutputs.length}) doesn't match external + change output length (${expectedOutputCount})`
-      )
-    }
+  //   if (externalOutputs.length !== data.externalOutputs.length) {
+  //     throw new Error(
+  //       `Invalid tx: externalOutputs length (${externalOutputs.length}) doesn't match data.externalOutputs length (${data.externalOutputs.length})`
+  //     )
+  //   }
+  //   const expectedOutputCount = data.externalOutputs.length + data.changeOutputs.length
+  //   if (psbtOutputs.length !== expectedOutputCount) {
+  //     throw new Error(
+  //       `Invalid tx: psbt outputs length (${psbtOutputs.length}) doesn't match external + change output length (${expectedOutputCount})`
+  //     )
+  //   }
 
-    if (externalOutputs.length === 1 && externalOutputs[0].address !== tx.toAddress) {
-      throw new Error(`Invalid tx: toAddress (${tx.toAddress}) doesn't match external output 0 address (${externalOutputs[0].address})`)
-    } else if (externalOutputs.length > 1 && tx.toAddress !== 'batch') {
-      throw new Error(`Invalid tx: toAddress (${tx.toAddress}) should be "batch" for multi output transaction`)
-    }
+  //   if (externalOutputs.length === 1 && externalOutputs[0].address !== tx.toAddress) {
+  //     throw new Error(`Invalid tx: toAddress (${tx.toAddress}) doesn't match external output 0 address (${externalOutputs[0].address})`)
+  //   } else if (externalOutputs.length > 1 && tx.toAddress !== 'batch') {
+  //     throw new Error(`Invalid tx: toAddress (${tx.toAddress}) should be "batch" for multi output transaction`)
+  //   }
 
-    let externalOutputTotal = new BigNumber(0) // main denom
-    let changeOutputTotal = new BigNumber(0) // main denom
+  //   let externalOutputTotal = new BigNumber(0) // main denom
+  //   let changeOutputTotal = new BigNumber(0) // main denom
 
-    // Safe to assume outputs are consistently ordered with external followed by change
-    for (let i = 0; i < psbtOutputs.length; i++) {
-      const psbtOutput = psbtOutputs[i]
-      if (i < externalOutputs.length) {
-        this.validatePsbtOutput(externalOutputs[i], psbtOutput, i)
-        this.validatePsbtOutput(data.externalOutputs[i], psbtOutput, i)
-        externalOutputTotal = externalOutputTotal.plus(data.externalOutputs[i].value)
-      } else {
-        const changeOutputIndex = i - externalOutputs.length
-        const changeOutput = data.changeOutputs[changeOutputIndex]
-        this.validatePsbtOutput(changeOutput, psbtOutput, i)
+  //   // Safe to assume outputs are consistently ordered with external followed by change
+  //   for (let i = 0; i < psbtOutputs.length; i++) {
+  //     const psbtOutput = psbtOutputs[i]
+  //     if (i < externalOutputs.length) {
+  //       this.validatePsbtOutput(externalOutputs[i], psbtOutput, i)
+  //       this.validatePsbtOutput(data.externalOutputs[i], psbtOutput, i)
+  //       externalOutputTotal = externalOutputTotal.plus(data.externalOutputs[i].value)
+  //     } else {
+  //       const changeOutputIndex = i - externalOutputs.length
+  //       const changeOutput = data.changeOutputs[changeOutputIndex]
+  //       this.validatePsbtOutput(changeOutput, psbtOutput, i)
 
-        // If we stop reusing addresses in the future this will need to be changed
-        if ((tx.fromAddress !== 'batch') && (changeOutput.address !== tx.fromAddress)) {
-          throw new Error(`Invalid tx: change output ${i} address (${changeOutput.address}) doesn't match fromAddress (${tx.fromAddress})`)
-        }
+  //       // If we stop reusing addresses in the future this will need to be changed
+  //       if ((tx.fromAddress !== 'batch') && (changeOutput.address !== tx.fromAddress)) {
+  //         throw new Error(`Invalid tx: change output ${i} address (${changeOutput.address}) doesn't match fromAddress (${tx.fromAddress})`)
+  //       }
 
-        if (data.changeAddress !== null && data.changeAddress !== changeOutput.address) {
-          throw new Error(`Invalid tx: change output ${i} address (${changeOutput.address}) doesn't match data.changeAddress (${data.changeAddress})`)
-        }
-        changeOutputTotal = changeOutputTotal.plus(changeOutput.value)
-      }
-    }
+  //       if (data.changeAddress !== null && data.changeAddress !== changeOutput.address) {
+  //         throw new Error(`Invalid tx: change output ${i} address (${changeOutput.address}) doesn't match data.changeAddress (${data.changeAddress})`)
+  //       }
+  //       changeOutputTotal = changeOutputTotal.plus(changeOutput.value)
+  //     }
+  //   }
 
-    // Check totals
+  //   // Check totals
 
 
-    if (data.inputTotal && !inputTotal.eq(data.inputTotal)) {
-      throw new Error(`Invalid tx: data.externalOutputTotal (${data.externalOutputTotal}) doesn't match expected external output total (${externalOutputTotal})`)
-    }
-    if (data.externalOutputTotal && !externalOutputTotal.eq(data.externalOutputTotal)) {
-      throw new Error(`Invalid tx: data.externalOutputTotal (${data.externalOutputTotal}) doesn't match expected external output total (${externalOutputTotal})`)
-    }
-    if (!changeOutputTotal.eq(data.change)) {
-      throw new Error(`Invalid tx: data.change (${data.externalOutputTotal}) doesn't match expected change output total (${externalOutputTotal})`)
-    }
-    if (!externalOutputTotal.eq(tx.amount)) {
-      throw new Error(`Invalid tx: amount (${tx.amount}) doesn't match expected external output total (${externalOutputTotal})`)
-    }
-    const expectedFee = inputTotal.minus(externalOutputTotal).minus(changeOutputTotal)
-    if (!expectedFee.eq(tx.fee)) {
-      throw new Error(`Invalid tx: fee (${tx.fee}) doesn't match expected fee (${expectedFee})`)
-    }
+  //   if (data.inputTotal && !inputTotal.eq(data.inputTotal)) {
+  //     throw new Error(`Invalid tx: data.externalOutputTotal (${data.externalOutputTotal}) doesn't match expected external output total (${externalOutputTotal})`)
+  //   }
+  //   if (data.externalOutputTotal && !externalOutputTotal.eq(data.externalOutputTotal)) {
+  //     throw new Error(`Invalid tx: data.externalOutputTotal (${data.externalOutputTotal}) doesn't match expected external output total (${externalOutputTotal})`)
+  //   }
+  //   if (!changeOutputTotal.eq(data.change)) {
+  //     throw new Error(`Invalid tx: data.change (${data.externalOutputTotal}) doesn't match expected change output total (${externalOutputTotal})`)
+  //   }
+  //   if (!externalOutputTotal.eq(tx.amount)) {
+  //     throw new Error(`Invalid tx: amount (${tx.amount}) doesn't match expected external output total (${externalOutputTotal})`)
+  //   }
+  //   const expectedFee = inputTotal.minus(externalOutputTotal).minus(changeOutputTotal)
+  //   if (!expectedFee.eq(tx.fee)) {
+  //     throw new Error(`Invalid tx: fee (${tx.fee}) doesn't match expected fee (${expectedFee})`)
+  //   }
 
-  }
+  // }
 
   validateAndFinalizeSignedTx(
     tx: BitcoinSignedTransaction | BitcoinUnsignedTransaction,
